melissa-api/pyproject.toml
[project]
name = "melissa-api"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "pydantic>=2.7",
  "httpx>=0.27",
  "python-multipart>=0.0.9",
  "PyNaCl>=1.5",
  "jsonschema>=4.23",
  "python-dotenv>=1.0.1",
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
app = "src.main:app"
reload = true
